+++ views/add-bookings.php
@@ -95,7 +95,7 @@ $pcm_courses_to_book = new WP_Query( apply_filters( 'pcm_courses_to_book_args', 
 
 <div class="wrap">
 
-	<h2><?php echo esc_html( get_admin_page_title() ); ?></h2>
+	<h1><?php echo esc_html( get_admin_page_title() ); ?></h1>
 
 	<?php if ( isset( $_GET['msg'] ) ) { ?>
 		<div id="message" class="updated">