+++ views/email-notifications.php
@@ -14,7 +14,7 @@ $pcm_email_notifications = get_option( 'pcm_email_notifications' );
 
 <div class="wrap">
 
-	<h2><?php echo esc_html( get_admin_page_title() ); ?></h2>
+	<h1><?php echo esc_html( get_admin_page_title() ); ?></h1>
 
 	<?php if ( isset( $_GET['msg'] ) ) { ?>
 		<div id="message" class="updated">