+++ CHANGES.md
@@ -1,6 +1,6 @@
 # Changelog for Pilau Course Management
 
-## 0.3.7
+## 0.3.7 (2015-07-31)
 * Improved styling for add bookings screen options
 * Added `pcm_admin_user_settings` filter
 * Added `pcm_manage_bookings_extra_cols_headings` and `pcm_manage_bookings_extra_cols_cells` actions